Rename Firebase utils to Firebase services
