+++ src/containers/specifics/archive/__tests__/Archives.test.tsx
@@ -1,5 +1,5 @@
 import Archives from '~/containers/specifics/archive/Archives';
-import { timestampFromDate } from '~/utils/firebase';
+import { timestampFromDate } from '~/services/firestore';
 import { mockStore, withProvider } from '~/utils/testHelpers';
 
 describe('ArchivesContainer', () => {
