+++ src/hooks/common/__tests__/useWorkout.test.ts
@@ -1,5 +1,5 @@
 import { useWorkout } from '~/hooks/common/useWorkout';
-import { timestampFromDate } from '~/utils/firebase';
+import { timestampFromDate } from '~/services/firestore';
 import { mockWorkouts } from '~/utils/mocks/mockWorkouts';
 import { mockStore, withHook } from '~/utils/testHelpers';
