+++ src/services/firestore/index.ts
@@ -1,0 +1,5 @@
+import { postError } from './postError';
+import { timestampFromDate } from './timestampFromDate';
+
+export * from './collections';
+export { postError, timestampFromDate };
