+++ src/services/firestore/timestampFromDate.ts
@@ -1,4 +1,4 @@
-import firebase, { Timestamp } from '~/utils/firebase';
+import firebase, { Timestamp } from '~/services/firebase';
 
 export const timestampFromDate = (date: Date): Timestamp =>
   firebase.firestore.Timestamp.fromDate(date);
