+++ src/utils/__tests__/datetime.test.ts
@@ -1,12 +1,12 @@
 import timekeeper from 'timekeeper';
 
+import { timestampFromDate } from '~/services/firestore';
 import {
   formatDayNumeric,
   formatShortMonth,
   formatUS,
   isToday,
 } from '~/utils/datetime';
-import { timestampFromDate } from '~/utils/firebase';
 
 describe('utils: datetime', () => {
   afterEach(() => {
