+++ src/utils/mocks/mockWorkouts.ts
@@ -1,5 +1,5 @@
+import { timestampFromDate } from '~/services/firestore';
 import { Workout } from '~/store/workout';
-import { timestampFromDate } from '~/utils/firebase';
 
 export const mockWorkouts = (): Workout[] => {
   const arr = Array.from(Array(30), (_v, k) => k + 1);
